refactor: improve monitoring services architecture and alert handling
- Refactored alert system in zeromqService to separate metrics collection from message formatting
- Improved monitoring service organization with dedicated modules for CPU, memory, and security monitoring
- Enhanced error handling and logging across monitoring services
- Implemented better state management for monitoring intervals
- Added graceful shutdown handling for monitoring processes
- Standardized alert cooldown periods and threshold management
- Improved type safety with better interfaces and error handling
