+++ package/package.json
@@ -1,6 +1,6 @@
 {
   "name": "telex-server-monitor-sdk",
-  "version": "1.0.20",
+  "version": "1.0.21",
   "publishConfig": {
     "access": "public"
   },
